backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
tensorflow==2.16.1
Pillow==10.0.1
PyJWT==2.8.0
Werkzeug==2.3.7
numpy>=1.26.0
h5py>=3.9.0
